changed package name, moved variable position, and other review changes
